requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.30.1
openai-whisper==20231117
requests==2.32.3
gunicorn==22.0.0
